Refactor parseRepo and generateSite into testable, concurrent modules
- Migrated general model structures (`SiteData`, `PackageData`, `VersionData`, `ManifestEntryData`, etc.) from `cmd/g2` to the root `g2` package.
- Extracted helper functions from the monolithic `parseRepo` into `cmd/g2/parse_repo.go`.
- Extracted and parallelized `generateSite` utilizing `errgroup.Group` for repository pages in `cmd/g2/generate_site.go`.
- Added programmatic `Context()` generation on `PageNode` to generate dynamic `BaseURL` and `Breadcrumbs` contexts, upgrading template interfaces.
