+++ cmd/g2/search_index.go
@@ -57,7 +57,7 @@ type SearchManifest struct {
 
 var pkgRegex = regexp.MustCompile(`([a-zA-Z0-9_][a-zA-Z0-9_\-\+]*\/[a-zA-Z0-9_][a-zA-Z0-9_\-\+]+)`)
 
-func generateSearchData(outDir, outZip string, sites []*SiteData, maxChunkSizeOverride ...int) error {
+func generateSearchData(outDir, outZip string, sites []*g2.SiteData, maxChunkSizeOverride ...int) error {
 	var documents []SearchDocument
 	docID := 0
 
@@ -387,7 +387,7 @@ func generateSearchData(outDir, outZip string, sites []*SiteData, maxChunkSizeOv
 	return nil
 }
 
-func generateSearchIndex(outDir string, sites []*SiteData) error {
+func generateSearchIndex(outDir string, sites []*g2.SiteData) error {
 	searchDir := filepath.Join(outDir, "search")
 	dataDir := filepath.Join(searchDir, "data")
