+++ cmd/g2/site_extra_test.go
@@ -1,16 +1,16 @@
 package main
 
 import (
-	"testing"
 	"github.com/arran4/g2"
+	"testing"
 )
 
 func TestResolveDependencies(t *testing.T) {
-	site := &SiteData{
-		Categories: []CategoryData{
+	site := &g2.SiteData{
+		Categories: []g2.CategoryData{
 			{
 				Name: "app-misc",
-				Packages: []PackageData{
+				Packages: []g2.PackageData{
 					{
 						Category: "app-misc",
 						Name:     "bar",
