+++ cmd/g2/uses_serve_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"github.com/arran4/g2"
 	"os"
 	"testing"
 	"time"
@@ -14,12 +15,12 @@ func TestGenerateUsesPages(t *testing.T) {
 
 	outDir := t.TempDir()
 
-	err = generateSite(outDir, []*SiteData{siteData}, 90*24*time.Hour, "3 months", GenerationInfo{})
+	err = generateSite(outDir, []*g2.SiteData{siteData}, 90*24*time.Hour, "3 months", GenerationInfo{})
 	if err != nil {
 		t.Fatalf("generateSite failed: %v", err)
 	}
 
-		if _, err := os.Stat(outDir + "/uses/index.html"); os.IsNotExist(err) {
+	if _, err := os.Stat(outDir + "/uses/index.html"); os.IsNotExist(err) {
 		t.Errorf("Expected /uses/index.html to be generated, got error: %v", err)
 	}
 }
