+++ templates/views/repo_index.html
@@ -152,11 +152,11 @@
     {{if gt (len .Repo.Authors) 0}}
     <li><a href="authors/">Authors ({{len .Repo.Authors}})</a></li>
     {{end}}
-    {{if gt (len .Repo.AggUseFlags) 0}}
-    <li><a href="uses/">USE Flags ({{len .Repo.AggUseFlags}})</a></li>
+    {{if gt (len_or_zero .Repo.AggUseFlags) 0}}
+    <li><a href="uses/">USE Flags ({{len_or_zero .Repo.AggUseFlags}})</a></li>
     {{end}}
-    {{if gt (len .Repo.AggEclasses) 0}}
-    <li><a href="eclasses/">Eclasses ({{len .Repo.AggEclasses}})</a></li>
+    {{if gt (len_or_zero .Repo.AggEclasses) 0}}
+    <li><a href="eclasses/">Eclasses ({{len_or_zero .Repo.AggEclasses}})</a></li>
     {{end}}
     {{if gt (len .Repo.InfoPkgs) 0}}
     <li><a href="info_pkgs/">Info Packages ({{len .Repo.InfoPkgs}})</a></li>
